Dialog refactoring complete
