--- src/dialogs/InputPublicKey/index.js
@@ -1,2 +1,0 @@
-import InputPublicKey from './InputPublicKey';
-export default InputPublicKey;
