--- src/dialogs/Receive/index.js
@@ -1,3 +1,0 @@
-import Receive from './Receive';
-
-export default Receive;
