--- src/dialogs/Send/index.js
@@ -1,2 +1,0 @@
-import Send from './Send';
-export default Send;
