--- src/dialogs/Send/styles.js
@@ -1,37 +1,0 @@
-import { StyleSheet } from 'react-native';
-import { colors, fontWeight } from '../../config/styling';
-import styleMerge from '../../utils/styleMerge';
-import parentStyles from '../dialogStyles';
-
-export default styleMerge(
-  parentStyles('light'),
-  StyleSheet.create({
-    container: {
-      paddingTop: 8,
-    },
-    deleteIcon: {
-      color: '#FFFFFF',
-      fontSize: 21,
-    },
-    deleteIconContainer: {
-      // marginLeft: -10
-    },
-    currencyButton: {
-      position: 'absolute',
-      right: 0,
-      bottom: 0,
-      borderWidth: 1,
-      borderColor: colors.purple,
-      borderRadius: 8,
-      paddingTop: 6,
-      paddingBottom: 5,
-      paddingHorizontal: 7,
-      marginBottom: 8,
-    },
-    currencyButtonText: {
-      color: '#617AF7',
-      ...fontWeight.medium,
-      letterSpacing: 0.1,
-    },
-  }),
-);
