--- src/dialogs/Sign/index.js
@@ -1,2 +1,0 @@
-import Sign from './Sign';
-export default Sign;
