+++ src/dialogs/SweepKeyDetails.js
@@ -1,6 +1,8 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import { FlatList, View, Alert } from 'react-native';
+import {
+  StyleSheet, FlatList, View, Alert,
+} from 'react-native';
 import Big from 'big.js';
 import {
   Button,
@@ -12,13 +14,85 @@ import {
   FeeSlider,
   ExpandableView,
   FontScale,
-} from '../../components';
-import styles from './styles';
-import { numFormat } from '../../utils/numFormat';
-import ExchangeHelper from '../../utils/exchangeHelper';
-import SettingHelper from '../../utils/settingHelper';
-import { getByteCount } from '../../libs/coinid-public/utils';
-import { fontSize, gridMultiplier } from '../../config/styling';
+} from '../components';
+import { numFormat } from '../utils/numFormat';
+import ExchangeHelper from '../utils/exchangeHelper';
+import SettingHelper from '../utils/settingHelper';
+import { getByteCount } from '../libs/coinid-public/utils';
+import {
+  colors, fontWeight, fontSize, gridMultiplier,
+} from '../config/styling';
+
+import styleMerge from '../utils/styleMerge';
+import parentStyles from './styles/common';
+
+const styles = styleMerge(
+  parentStyles('light'),
+  StyleSheet.create({
+    secondTitle: {
+      textAlign: 'center',
+      fontSize: fontSize.h3,
+      lineHeight: fontSize.h3 * 1.2,
+      marginBottom: gridMultiplier * 2,
+      ...fontWeight.medium,
+    },
+    balance: {
+      marginBottom: gridMultiplier * 1,
+      textAlign: 'center',
+      color: colors.purple,
+      fontSize: fontSize.large,
+      lineHeight: fontSize.large * 1.2,
+      ...fontWeight.bold,
+    },
+    fiatBalance: {
+      marginBottom: gridMultiplier * 2 - 2,
+      textAlign: 'center',
+      color: colors.gray,
+      fontSize: fontSize.h2,
+      lineHeight: fontSize.large * 1.2,
+      ...fontWeight.normal,
+    },
+    detailItem: {
+      paddingVertical: gridMultiplier * 1,
+      paddingHorizontal: gridMultiplier * 2,
+    },
+    detailItemFirst: {
+      paddingTop: gridMultiplier * 1,
+    },
+    detailItemBalance: {
+      fontSize: fontSize.base,
+      lineHeight: fontSize.base * 1.2,
+      ...fontWeight.medium,
+      marginBottom: gridMultiplier * 1 - 2,
+    },
+    feeWrapper: {
+      marginBottom: gridMultiplier * 2,
+    },
+    totalWrapper: {
+      marginBottom: gridMultiplier * 3,
+    },
+    detailItemAddress: {
+      fontSize: fontSize.small,
+      lineHeight: fontSize.small * 1.2,
+    },
+    addressContainerWrapper: {
+      marginHorizontal: -gridMultiplier * 2,
+      marginTop: gridMultiplier * 1,
+    },
+    addressContainer: {},
+    addressContainerContent: {
+      marginTop: -gridMultiplier * 1,
+      marginBottom: gridMultiplier * 1,
+    },
+    horizontalBorder: {
+      height: 1,
+      marginBottom: gridMultiplier * 3,
+      paddingHorizontal: gridMultiplier * 2,
+      marginHorizontal: -gridMultiplier * 2,
+      backgroundColor: colors.lightGray,
+    },
+  }),
+);
 
 export default class SweepKeyDetails extends PureComponent {
   constructor(props, context) {
