--- src/dialogs/SweepKeyDetails/index.js
@@ -1,3 +1,0 @@
-import SweepKeyDetails from './SweepKeyDetails';
-
-export default SweepKeyDetails;
