--- src/dialogs/TransactionDetails/index.js
@@ -1,2 +1,0 @@
-import TransactionDetails from './TransactionDetails';
-export default TransactionDetails;
